+++ official/projects/roformer/roformer_encoder_test.py
@@ -19,7 +19,7 @@ import numpy as np
 import tensorflow as tf
 
 from tensorflow.python.keras import keras_parameterized  # pylint: disable=g-direct-tensorflow-import
-from official.nlp.modeling.networks import roformer_encoder
+import roformer_encoder
 
 
 # This decorator runs the test in V1, V2-Eager, and V2-Functional mode. It
